LayoutTests/crypto/workers/subtle/hmac-import-raw-key-expected.txt

[Worker] Test importing a raw HMAC key with SHA-1 in workers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/hmac-import-raw-key.js
[Worker] Importing a key...
PASS [Worker] crypto.subtle.importKey("raw", asciiToUint8Array(""), {name: "hmac", hash: "sha-1"}, extractable, ["sign", "verify"]) rejected promise  with DataError (DOM Exception 30): Data provided to an operation does not meet requirements.
PASS [Worker] key.type is 'secret'
PASS [Worker] key.extractable is true
PASS [Worker] key.algorithm.name is 'HMAC'
PASS [Worker] key.algorithm.length is 128
PASS [Worker] key.algorithm.hash.name is 'SHA-1'
PASS [Worker] key.usages is ['sign', 'verify']
PASS successfullyParsed is true

TEST COMPLETE
